deleteMultipleImages stores failed images deletion in schema

+++ server/src/utils/cloudinaryUploader.js
@@ -1,18 +1,24 @@
+const FailedDeletion = require('../models/failedImgDeletionSchema');
+
 const streamifier = require('streamifier');
 const { cloudinary } = require('../config/cloudinary');
 const { fixInvalidFolderNameChars } = require('./utils');
 const { imagesOptions } = require('../config/cloudinary');
 const { createValidationError } = require('./createValidationError');
 const { errorMessages } = require('../constants/errorMessages');
+const { validateImages } = require('./validateImages');
 
-async function addImages(images, obj, folderName) {
+async function addImages(images, data, folderName, minImagesRequired = 1) {
+    // Validating the image files with default minimum number of images required
+    validateImages(images, minImagesRequired);
+
     const imageUrls = [];
     let imgError = null;
 
     try {
         // Set the folder type and name for Cloudinary
         const folder_type = folderName;
-        const folder_name = fixInvalidFolderNameChars(obj.city, obj._id);
+        const folder_name = fixInvalidFolderNameChars(data.city, data._id);
 
         // Upload the images to Cloudinary
         const cloudinaryImagesData = await handleImageUploads(
@@ -84,6 +90,7 @@ async function handleImageUploads(files, options = {}) {
         console.error('Error uploading image:', error);
     });
 
+    // ! Add an error message to notify the client if there is some failed uploads
     return fulfilledResults;
 }
 
@@ -109,28 +116,97 @@ function uploadImageToCloudinary(imageBuffer, options = {}) {
 }
 
 async function deleteMultipleImages(public_ids, folderNames) {
+    // Validate the props
     if (!Array.isArray(public_ids) || !Array.isArray(folderNames)) {
         throw createValidationError(errorMessages.cloudinaryValidation, 400);
     }
 
     try {
-        const promises_ids = public_ids.map((x) => deleteImage(x));
-        await Promise.all(promises_ids);
+        const promises_ids = public_ids.map((id) => deleteImage(id));
+        const results = await Promise.allSettled(promises_ids);
 
+        // Filters out the rejected promises
+        let failedPromises = results.filter((result) => result.status === 'rejected');
+
+        const failedToDeleteAfterRetry = await retryDeletion(failedPromises);
+
+        // If there is still undeleted images - it stores the public ids (to delete them later)
+        if (failedToDeleteAfterRetry.length > 0) {
+            const failedPublicIds = failedToDeleteAfterRetry.map(
+                (result) => result.reason.publicId
+            );
+
+            // Stores the ids in the specified mongoDB schema
+            FailedDeletion.create({ public_ids: failedPublicIds }).catch((err) =>
+                console.error(err.message)
+            );
+
+            // Calculate how many images have been deleted
+            const deletedImgs = promises_ids.length - failedToDeleteAfterRetry.length;
+            const initialImages = promises_ids.length;
+
+            // Log the result
+            const errorMessage = `Deleted: ${deletedImgs} out of ${initialImages}`
+            
+            throw new Error(errorMessage);
+        }
+
+
+        // Delete the folder/folders upon success
         folderNames.forEach((folder) => cloudinary.api.delete_folder(folder));
 
         return true;
     } catch (error) {
-        console.log(error.message);
         throw error;
+    }
+
+    async function retryDeletion(promises) {
+        let retryAttempts = 0;
+        const MAX_RETRY_ATTEMPTS = 2;
+
+        // Copy array
+        let failedPromises = promises.slice();
+
+        // Retry the deletion of failed images when there's rejected promises
+        while (retryAttempts < MAX_RETRY_ATTEMPTS && failedPromises.length > 0) {
+            const retryPromises = failedPromises.map((result) =>
+                deleteImage(result.reason.publicId, retryAttempts)
+            );
+
+            const retryResults = await Promise.allSettled(retryPromises);
+
+            // Filter out still failed promises for the next iteration
+            failedPromises = retryResults.filter((result) => result.status === 'rejected');
+
+            retryAttempts++;
+
+            // 1s delay before the next retry
+            await delay(1000);
+        }
+
+        return failedPromises;
+    }
+
+    // Utility function to introduce a delay
+    function delay(ms) {
+        return new Promise((resolve) => setTimeout(resolve, ms));
     }
 }
 
 async function deleteImage(publicId) {
-    const res = await cloudinary.uploader.destroy(publicId);
+    try {
+        const res = await cloudinary.uploader.destroy(publicId);
 
-    if (res.result !== 'ok') {
-        throw createValidationError(errorMessages.notFound, 404);
+        if (res.result !== 'ok') {
+            console.log(`deleteImage: ${res}`);
+            throw new Error();
+        }
+
+        return true;
+    } catch (err) {
+        console.log(err?.message || '')
+        err.publicId = publicId;
+        throw err;
     }
 }
